+++ src/apprc/__init__.py
@@ -7,6 +7,7 @@ from apprc.public.field import field
 from . import cli
 from . import files
 from . import provenance
+from . import schema
 from . import storage
 
 __all__ = [
@@ -17,5 +18,6 @@ __all__ = [
     "cli",
     "files",
     "provenance",
+    "schema",
     "storage",
 ]